requirements.txt
# Environment & Configuration
python-dotenv>=1.0.0

# Reddit Collection
praw>=7.7.0

# YouTube Collection
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Hugging Face Datasets
datasets>=2.14.0
huggingface-hub>=0.17.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Language Detection
langdetect>=1.0.9

# Data Versioning
dvc>=3.0.0

# General
requests>=2.31.0
tqdm>=4.65.0
